requirements.txt
django==2.2.12
django_cors_headers
graphene_django==2.8.2
pyjwt
requests